+++ src/hooks/useTransactionWithToast.tsx
@@ -54,8 +54,6 @@ export function useTransactionWithToast({
     [hash],
   );
 
-  console.log('isConfirming', isConfirming);
-
   useEffect(() => {
     if (isConfirming) {
       toast.loading(renderToastContent(pendingText, pendingDescription), {
